pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "github_tests_validator_app"
version = "0.1.0"
description = "`github_tests_validator_app` is a Python cli/package"
readme = "README.md"
authors = [
  "artefactory <hello@artefactory.com>"
]
license = "MIT"
repository = "https://github.com/artefactory/github_tests_validator_app"
homepage = "https://github.com/artefactory/github_tests_validator_app"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  # Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  # Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"github_tests_validator_app" = "github_tests_validator_app.cli.__main__:app"

[tool.poetry.dependencies]
python = "^3.9"
importlib_metadata = {version = ">=1.6.0", python = "<3.8"}
typer = {extras = ["all"], version = ">=0.3.2"}
rich = ">=10.1.0"
fastapi = ">=0.78.0"
uvicorn = ">=0.18.2"
PyJWT = ">=2.4.0"
requests = ">=2.22.0"
PyGithub = ">=1.55"
cryptography = ">=36.0.1"
urllib3 = ">=1.26.5"
gspread = "^5.4.0"
types-requests = "^2.28.9"
pytest-mock = "^3.8.2"
types-PyYAML = "^6.0.11"
google-api-python-client = "^2.60.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"
google-cloud-secret-manager = "^2.12.4"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
darglint = ">=1.8.0"
isort = ">=5.8.0"
pyupgrade = ">=2.12.0"
black = ">=20.8b1"
mypy = ">=0.812"
bandit = ">=1.7.0"
safety = ">=1.10.3"
pytest = ">=6.2.1"
pylint = ">=2.6.0"
pydocstyle = ">=6.0.0"
pre-commit = ">=2.12.0"
mypy-extensions = ">=0.4.3"

[tool.black]
# https://github.com/psf/black
line-length = 100
target-version = ["py38"]

[tool.isort]
# https://github.com/timothycrosley/isort/
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
indent = 4
force_grid_wrap = 0
use_parentheses = true
line_length = 100
